Dynamic Sizing forX Samples
So, instead of a fixed 3s or 5s sample I've reduced it to 140% of the total and that is all controlled with a simple number in the main script.

The model still isn't doing a great job, so need to move back to a single output (and non-categorical). Right now it still just predicts ALL one thing or all the other.